requirements.txt
Flask
Flask-Login
mysql-connector-python
Werkzeug
python-dotenv